meme/requirements.txt
asgiref==3.2.7
astroid==2.3.3
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.2
django-crispy-forms==1.9.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
path==13.1.0
Pillow==7.0.0
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
